reafactor(CarsController): se limpia codigo de trycatch

+++ app/Controllers/Http/CarsController.ts
@@ -1,5 +1,6 @@
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import { CreateCarDTO } from 'App/DTOs/Car/createCarDTOs';
+import NotFoundException from 'App/Exceptions/NotFoundException';
 
 import Car from "App/Models/Car";
 import CarService from 'App/Services/CarService';
@@ -7,32 +8,23 @@ import CarValidator from 'App/Validators/CarValidator';
 
 export default class CarsController {
     public async findAll({ response }: HttpContextContract) {
-        try {
-            const cars = await Car.all();
-            return response.status(200).send(cars);
-        } catch (error) {
-            console.log(error);
-            return response.status(500).send('Error fetching cars');
-        }
+        const cars = await Car.all();
+        return response.ok(cars);
     }
     public async delete({ params, response }: HttpContextContract) {
-        try {
-            const car = await Car.findOrFail(params.id);
-            await car.delete();
-            return response.status(204).send('Car deleted successfully');
-        } catch (error) {
-            console.log(error);
-            return response.status(500).send('Error deleting car');
+        const car = await Car.findOrFail(params.id);
+        if (!car) {
+            throw new NotFoundException('Car not found');
         }
+        await car.delete();
+        return response.ok({
+            status: 'success',
+            message: 'Car deleted successfully',
+        })
     }
     public async create({ request, response }: HttpContextContract) {
         const body: CreateCarDTO = await request.validate(CarValidator);
-        try {
-            const createdCar = await CarService.createCar(body);
-            return response.status(201).send(createdCar);
-        } catch (error) {
-            console.log(error);
-            return response.status(500).send('Error creating car');
-        }
+        const createdCar = await CarService.createCar(body);
+        return response.created(createdCar);
     }
 }
